Alpaca Algo Trading/requirements.txt
alpaca-py>=0.14.0
pandas>=2.0.0
tomli>=2.0.0; python_version < "3.11"
